Move `battery_pool.SystemBounds` to `timeseries._base_types`
This allows the type to be used by the other component pools and
reduces the PowerManagingActor's dependence on the battery_pool
package.
