+++ src/frequenz/sdk/timeseries/_base_types.py
@@ -3,6 +3,7 @@
 
 """Timeseries basic types."""
 
+import dataclasses
 import functools
 import typing
 from collections.abc import Callable, Iterator
@@ -10,7 +11,7 @@ from dataclasses import dataclass
 from datetime import datetime, timezone
 from typing import Generic, Self, overload
 
-from ._quantities import QuantityT
+from ._quantities import Power, QuantityT
 
 UNIX_EPOCH = datetime.fromtimestamp(0.0, tz=timezone.utc)
 """The UNIX epoch (in UTC)."""
@@ -150,3 +151,28 @@ class Bounds(Generic[_T]):
 
     upper: _T
     """Upper bound."""
+
+
+@dataclass(frozen=True)
+class SystemBounds:
+    """Internal representation of system bounds for groups of components."""
+
+    # compare = False tells the dataclass to not use name for comparison methods
+    timestamp: datetime = dataclasses.field(compare=False)
+    """Timestamp of the metrics."""
+
+    inclusion_bounds: Bounds[Power] | None
+    """Total inclusion power bounds for all components of a pool.
+
+    This is the range within which power requests would be allowed by the pool.
+
+    When exclusion bounds are present, they will exclude a subset of the inclusion
+    bounds.
+    """
+
+    exclusion_bounds: Bounds[Power] | None
+    """Total exclusion power bounds for all components of a pool.
+
+    This is the range within which power requests are NOT allowed by the pool.
+    If present, they will be a subset of the inclusion bounds.
+    """
