+++ src/frequenz/sdk/timeseries/battery_pool/_battery_pool.py
@@ -17,6 +17,7 @@ from ... import timeseries
 from ..._internal._channels import ReceiverFetcher
 from ...actor import _power_managing
 from ...timeseries import Energy, Percentage, Power, Sample, Temperature
+from .._base_types import SystemBounds
 from ..formula_engine import FormulaEngine
 from ..formula_engine._formula_generators import (
     BatteryPowerFormula,
@@ -30,7 +31,6 @@ from ._metric_calculator import (
     SoCCalculator,
     TemperatureCalculator,
 )
-from ._result_types import SystemBounds
 
 # pylint: disable=protected-access
